nfs: Fix shifting left
Shifting signed 0xffffffff 32 bit left is undefined behavior. Cast to
unsigned to make it defined behavior.

+++ nfs/nfs.c
@@ -401,7 +401,7 @@ xdr_decode_64bit (int *p, long long *n)
   p++;
   low = ntohl (*p);
   p++;
-  *n = ((high & 0xffffffff) << 32) | (low & 0xffffffff);
+  *n = (((uint64_t)(high & 0xffffffff)) << 32) | (low & 0xffffffff);
   return p;
 }
